+++ CHANGELOG.rst
@@ -9,7 +9,8 @@ Change Log
 0.8.3
 =====
 
-* 2024-05-??/dmichaels/PR-7
+* 2024-05-14/dmichaels/PR-10
+* Added rclone support.
 * Added metadata_template.py module with goal of checking the user's metadata
   file with the latest HMS DBMI metadata template and giving a warning if the
   version appears to be out of date. Also new convenience command to export and