+++ CHANGELOG.rst
@@ -6,6 +6,12 @@ smaht-submitr
 Change Log
 ----------
 
+0.8.4
+=====
+* Experimenting with pyinstaller to create a single independent executable (per command)
+  so commands can be run without having Python/pyenv/etc installed.
+
+
 0.8.3
 =====